Split Config.php
