+++ process/Model/Config/Config.php
@@ -1,47 +1,19 @@
 <?php
 
-namespace App\Model;
-
-class ExcludeRelationWay
-{
-  /** @var int[] */
-    public array $relation;
-  /** @var int[] */
-    public array $way;
-
-    public function __construct(array $relation, array $way)
-    {
-        $this->relation = $relation;
-        $this->way = $way;
-    }
-}
-
-class GenderRelationWay
-{
-  /** @var array<string,string> */
-    public array $relation;
-  /** @var array<string,string> */
-    public array $way;
-
-    public function __construct(array $relation, array $way)
-    {
-        $this->relation = $relation;
-        $this->way = $way;
-    }
-}
+namespace App\Model\Config;
 
 class Config
 {
     public int $relationId;
 
-  /** @var string[] */
+    /** @var string[] */
     public array $languages;
 
     public ExcludeRelationWay $exclude;
 
     public GenderRelationWay $gender;
 
-  /** @var array<string,bool> */
+    /** @var array<string,bool> */
     public array $instances;
 
     public function __construct(array $config)
